Develop (#11)
* update company api

+++ app/Http/Controllers/CompanyController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Company;
+use App\Models\User;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Resources\CompanyResource;
 use App\Traits\UploadTrait;
@@ -22,13 +23,18 @@ class CompanyController extends Controller
     public function index($user_id)
     {
         try {
+            $user = User::find( $user_id );
+
+            if( !$user ) {
+                return response()->json('User Not Found');
+            }
+
             $companies = Company::where( 'users_id', $user_id )->orderBy('id', 'DESC')->get();
         } catch (\Exception $ex) {
             return response()->json(config('naz.db'), config('naz.db_error'));
         }
 
         return CompanyResource::collection($companies);
-        //return response()->json($areas);
     }
 
     /**
@@ -40,22 +46,33 @@ class CompanyController extends Controller
     public function store( $user_id, Request $request )
     {
         $validator = Validator::make($request->all(), [
-            'name' => 'required|string|min:3|unique:Areas,name',
-            'status' => 'required|integer'
+            'name'          => 'required|string|min:3|unique:companies,name',
+            'status'        => 'required|boolean',
+            'logo'          => 'sometimes|nullable|array',
+            'description'   => 'sometimes|nullable|array',
+            'contact'       => 'sometimes|nullable|max:15|string',
+            'contact_person'=> 'sometimes|nullable|string',
+            'website'       => 'sometimes|nullable|string'
         ]);
 
         if ($validator->fails()) return response()->json($validator->errors(), config('naz.validation'));
 
         try {
+            $user = User::find( $user_id );
+
+            if( !$user ) {
+                return response()->json('User Not Found');
+            }
+
             $company = new Company();
             
-            $company->name     = $request->name;
-            $company->status   = $request->status;
-            $company->users_id = $user_id;
-            
-            if( isset( $request->description ) ) {
-                $company->description = $request->description;
-            }
+            $company->name           = $request->name;
+            $company->status         = $request->status;
+            $company->users_id       = $user_id;
+            $company->description    = $request->description;
+            $company->contact        = $request->contact;
+            $company->contact_person =  $request->contact_person; 
+            $company->website        = $request->website;
 
             if ( $request->has('logo') ) {
                 // Get image file
@@ -63,7 +80,7 @@ class CompanyController extends Controller
                 // Make a image name based on user name and current timestamp
                 $name = Str::slug($request->input('name')) . '_' . time();
                 // Define folder path
-                $folder = '/uploads/categories/';
+                $folder = '/uploads/company/';
                 // Make a file path where image will be stored [ folder path + file name + file extension]
                 $filePath = $folder . $name . '.' . $image->getClientOriginalExtension();
                 // Upload image
@@ -72,18 +89,6 @@ class CompanyController extends Controller
                 $company->logo = $filePath;
             }
 
-            if( isset( $request->contact ) ) {
-                $company->contact = $request->contact; 
-            }
-            
-            if( isset( $request->contact_person ) ) {
-                $company->contact_person =  $request->contact_person; 
-            }
-
-            if( isset( $request->contact_person ) ) {
-                $company->contact_person = $request->contact_person;
-            }
-        
             $company->save();
 
         } catch (\Exception $ex) {
@@ -104,6 +109,12 @@ class CompanyController extends Controller
     public function show( $user_id, $id )
     {
         try{
+            $user = User::find( $user_id );
+
+            if( !$user ) {
+                return response()->json('User Not Found');
+            }
+
             $company = Company::where( 'users_id', $user_id )->find($id);
 
         } catch (\Exception $ex) {
@@ -125,22 +136,34 @@ class CompanyController extends Controller
     {
         
         $validator = Validator::make($request->all(), [
-            'name' => 'required|string|min:3|unique:Areas,name',
+            'name'          => 'required|string|min:3|unique:companies,name',
+            'status'        => 'required|boolean',
+            'logo'          => 'sometimes|nullable|array',
+            'description'   => 'sometimes|nullable|array',
+            'contact'       => 'sometimes|nullable|max:15|string',
+            'contact_person'=> 'sometimes|nullable|string',
+            'website'       => 'sometimes|nullable|string'
         ]);
 
         if ( $validator->fails() ) return response()->json( $validator->errors(), config('naz.validation') );
 
         try {
 
+            $user = User::find( $user_id );
+
+            if( !$user ) {
+                return response()->json('User Not Found');
+            }
+
             $company = Company::where( 'users_id', $user_id )->find($id);
           
-            $company->name     = $request->name;
-            $company->status   = $request->status;
-            $company->users_id = $user_id;
-            
-            if( isset( $request->description ) ) {
-                $company->description = $request->description;
-            }
+            $company->name           = $request->name;
+            $company->status         = $request->status;
+            $company->users_id       = $user_id;
+            $company->description    = $request->description;
+            $company->contact        = $request->contact;
+            $company->contact_person = $request->contact_person;
+            $company->website        = $request->website;
 
             if ( $request->has('logo') ) {
                 // Get image file
@@ -156,19 +179,7 @@ class CompanyController extends Controller
                 // Set user profile image path in database to filePath
                 $company->logo = $filePath;
             }
-
-            if( isset( $request->contact ) ) {
-                $company->contact = $request->contact; 
-            }
             
-            if( isset( $request->contact_person ) ) {
-                $company->contact_person =  $request->contact_person; 
-            }
-
-            if( isset( $request->contact_person ) ) {
-                $company->contact_person = $request->contact_person;
-            }
-        
             $company->save();
 
         } catch (\Exception $ex) {
@@ -189,6 +200,12 @@ class CompanyController extends Controller
     public function destroy( $user_id, $id )
     {
         try{
+            $user = User::find( $user_id );
+
+            if( !$user ) {
+                return response()->json('User Not Found');
+            }
+            
             Company::where( 'users_id', $user_id )->where('id', $id)->delete();
         } catch (\Exception $ex) {
             return response()->json(config('naz.db'), config('naz.db_error'));
